old-tests/submission/Opera/script_scheduling/087.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: multiple defer scripts, one slow loading</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script src="scripts/delay.php?time=2&url=include-1.js" defer></script>
	<script src="scripts/include-2.js" defer></script>

	<script type="text/javascript">
        var t = async_test();
	function test() {
                assert_array_equals(eventOrder, ['external script #1', 'external script #2']);
		t.done();
        }
        onload = t.step_func(test);
	</script>

</body></html>
